Updated login module
